refactor: consolidate object store factory implementation
